add test for existence of backup directory

+++ backup_rsync.sh
@@ -1,6 +1,8 @@
 #!/bin/sh
 ip=192.168.1.164
-cd /travel/backup/renee
+backup_dir=/travel/backup/renee
+test -d $backup_dir || exit 1
+cd $backup_dir
 rm -rf backup.3
 mv backup.2 backup.3
 mv backup.1 backup.2
